refactor(api/common): move BackgroundServices to HostedServices
